+++ src/hint_processor/cairo_1_hint_processor/hint_processor.rs
@@ -4,10 +4,9 @@ use crate::any_box;
 use crate::felt::{felt_str, Felt252};
 use crate::hint_processor::cairo_1_hint_processor::dict_manager::DictSquashExecScope;
 use crate::hint_processor::hint_processor_definition::HintReference;
-use crate::types::relocatable::Relocatable;
-
 use crate::stdlib::collections::HashMap;
 use crate::stdlib::prelude::*;
+use crate::types::relocatable::Relocatable;
 use crate::{
     hint_processor::hint_processor_definition::HintProcessor,
     types::exec_scope::ExecutionScopes,
@@ -27,6 +26,7 @@ use core::ops::Mul;
 use num_bigint::BigUint;
 use num_integer::Integer;
 use num_traits::{cast::ToPrimitive, Zero};
+use std::ops::Shl;
 
 /// Execution scope for constant memory allocation.
 struct MemoryExecScope {
@@ -232,6 +232,25 @@ impl Cairo1HintProcessor {
                 high,
                 low,
             })) => self.wide_mul_128(vm, lhs, rhs, high, low),
+
+            Hint::Core(CoreHintBase::Core(CoreHint::Uint512DivModByUint256 {
+                dividend0,
+                dividend1,
+                dividend2,
+                dividend3,
+                divisor0,
+                divisor1,
+                quotient0,
+                quotient1,
+                quotient2,
+                quotient3,
+                remainder0,
+                remainder1,
+            })) => self.uint512_div_mod(
+                vm, dividend0, dividend1, dividend2, dividend3, divisor0, divisor1, quotient0,
+                quotient1, quotient2, quotient3, remainder0, remainder1,
+            ),
+
             hint => Err(HintError::UnknownHint(hint.to_string())),
         }
     }
@@ -597,6 +616,70 @@ impl Cairo1HintProcessor {
 
         vm.insert_value((dict_address + 1)?, prev_value)?;
         dict_manager_exec_scope.insert_to_tracker(dict_address, key, value);
+        Ok(())
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    fn uint512_div_mod(
+        &self,
+        vm: &mut VirtualMachine,
+        dividend0: &ResOperand,
+        dividend1: &ResOperand,
+        dividend2: &ResOperand,
+        dividend3: &ResOperand,
+        divisor0: &ResOperand,
+        divisor1: &ResOperand,
+        quotient0: &CellRef,
+        quotient1: &CellRef,
+        quotient2: &CellRef,
+        quotient3: &CellRef,
+        remainder0: &CellRef,
+        remainder1: &CellRef,
+    ) -> Result<(), HintError> {
+        let pow_2_128 = BigUint::from(u128::MAX) + 1u32;
+        let dividend0 = res_operand_get_val(vm, dividend0)?.to_biguint();
+        let dividend1 = res_operand_get_val(vm, dividend1)?.to_biguint();
+        let dividend2 = res_operand_get_val(vm, dividend2)?.to_biguint();
+        let dividend3 = res_operand_get_val(vm, dividend3)?.to_biguint();
+        let divisor0 = res_operand_get_val(vm, divisor0)?.to_biguint();
+        let divisor1 = res_operand_get_val(vm, divisor1)?.to_biguint();
+        let dividend: BigUint =
+            dividend0 + dividend1.shl(128) + dividend2.shl(256) + dividend3.shl(384);
+        let divisor = divisor0 + divisor1.shl(128);
+        let (quotient, remainder) = dividend.div_rem(&divisor);
+        let (quotient, limb0) = quotient.div_rem(&pow_2_128);
+
+        vm.insert_value(
+            cell_ref_to_relocatable(quotient0, vm)?,
+            Felt252::from(limb0),
+        )?;
+
+        let (quotient, limb1) = quotient.div_rem(&pow_2_128);
+        vm.insert_value(
+            cell_ref_to_relocatable(quotient1, vm)?,
+            Felt252::from(limb1),
+        )?;
+        let (limb3, limb2) = quotient.div_rem(&pow_2_128);
+
+        vm.insert_value(
+            cell_ref_to_relocatable(quotient2, vm)?,
+            Felt252::from(limb2),
+        )?;
+        vm.insert_value(
+            cell_ref_to_relocatable(quotient3, vm)?,
+            Felt252::from(limb3),
+        )?;
+        let (limb1, limb0) = remainder.div_rem(&pow_2_128);
+
+        vm.insert_value(
+            cell_ref_to_relocatable(remainder0, vm)?,
+            Felt252::from(limb0),
+        )?;
+        vm.insert_value(
+            cell_ref_to_relocatable(remainder1, vm)?,
+            Felt252::from(limb1),
+        )?;
+
         Ok(())
     }